Replace ClientBasicAuthenticationFilter.
The UAA (and spring) do not properly handle the decoding of client id
and secret when supplied as a basic auth header. The id and secret
should have been url encoded but the UAA was not handling this subtlety.

Before this commit, this filter extended the BasicAuthentication filter
in order to apply some preprocessing. As far as we can tell, this
preprocessing is no longer relevant.

Now, we implement a new version of the BasicAuthentication filter which
correctly handles the url decoding.

[#166970393]

Signed-off-by: Andrew Wittrock <awittrock@pivotal.io>